#41397 #41350 #41351

+++ app/src/main/java/com/timelysoft/tsjdomcom/service/model/payment/InvoicesIssuedModel.kt
@@ -10,7 +10,7 @@ class InvoicesIssuedModel (
 
     @SerializedName("Sum")
     @Expose
-    var sum: Int = 0,
+    var sum: Double = 0.0,
 
     @SerializedName("OnDate")
     @Expose
